# frontend/openapi.yaml
openapi: 3.0.0
info:
  title: Calendar Booking API
  version: 0.0.0
tags: []
paths:
  /admin/bookings:
    get:
      operationId: AdminBookings_list
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
  /admin/bookings/{id}:
    put:
      operationId: AdminBookings_update
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBookingRequest'
    delete:
      operationId: AdminBookings_delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
  /admin/event-types:
    post:
      operationId: AdminEventTypes_create
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventTypeRequest'
  /admin/event-types/{id}:
    put:
      operationId: AdminEventTypes_update
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventTypeRequest'
    delete:
      operationId: AdminEventTypes_delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
  /bookings:
    post:
      operationId: PublicBookings_create
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
  /event-types:
    get:
      operationId: PublicEventTypes_list
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventType'
  /event-types/{eventTypeId}/slots:
    get:
      operationId: PublicEventTypes_getAvailableSlots
      parameters:
        - name: eventTypeId
          in: path
          required: true
          schema:
            type: string
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date-time
          explode: false
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date-time
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Slot'
  /event-types/{id}:
    get:
      operationId: PublicEventTypes_getById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
components:
  schemas:
    Booking:
      type: object
      required:
        - id
        - eventTypeId
        - guestName
        - guestEmail
        - startTime
        - endTime
        - createdAt
      properties:
        id:
          type: string
        eventTypeId:
          type: string
        guestName:
          type: string
        guestEmail:
          type: string
          format: email
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    CreateBookingRequest:
      type: object
      required:
        - eventTypeId
        - guestName
        - guestEmail
        - startTime
      properties:
        eventTypeId:
          type: string
        guestName:
          type: string
        guestEmail:
          type: string
          format: email
        startTime:
          type: string
          format: date-time
    CreateEventTypeRequest:
      type: object
      required:
        - title
        - description
        - durationMinutes
      properties:
        title:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
          format: int32
          minimum: 1
    ErrorResponse:
      type: object
      required:
        - message
        - code
      properties:
        message:
          type: string
        code:
          type: string
    EventType:
      type: object
      required:
        - id
        - title
        - description
        - durationMinutes
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
          format: int32
          minimum: 1
    Slot:
      type: object
      required:
        - startTime
        - endTime
      properties:
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
    UpdateBookingRequest:
      type: object
      properties:
        guestName:
          type: string
        guestEmail:
          type: string
          format: email
        startTime:
          type: string
          format: date-time
    UpdateEventTypeRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
          format: int32
          minimum: 1